backend/requirements.txt
flask
flask-cors
openai-whisper
torch
torchaudio
pydub
numpy
